+++ packages/codenic_exception_converter/lib/src/exception_converter_suite.dart
@@ -75,81 +75,76 @@ class ExceptionConverterSuite {
   ///
   /// If no error occurs, then value [T] is returned.
   ///
-  /// If an [Exception] is thrown, then it is handled by the
-  /// [exceptionConverters] and [ExceptionConverterSuite.exceptionConverters].
+  /// If an [Exception] is thrown, then it will be converted into a [Failure]
+  /// using the [ExceptionConverter]s provided in the [exceptionConverters]
+  /// and [ExceptionConverterSuite.exceptionConverters], if any. If no matching
+  /// [ExceptionConverter] is found, then the [FallbackExceptionConverter] will
+  /// be used.
   ///
-  /// If the [Exception] does not have an exception converter, then it will be
-  /// converted to [Failure] by default using the [FallbackExceptionConverter].
-  ///
-  /// NOTE:
-  /// -  [Error]s are not caught by this method considering that it is bad
-  /// practice to catch an [Error].
-  /// See https://groups.google.com/a/dartlang.org/g/misc/c/lx9CXiV3o30/m/s5l_PwpHUGAJ.
-  Future<Either<Failure, T>> observe<T>({
+  /// If an [Error] is thrown, then it will be rethrown or handled by the
+  /// [onError] function, if provided.
+  FutureOr<Either<Failure, T>> observe<T>({
     required FutureOr<Either<Failure, T>> Function(MessageLog? messageLog) task,
     List<ExceptionConverter<Exception, T>>? exceptionConverters,
     MessageLog? messageLog,
     bool printOutput = false,
+    FutureOr<Either<Failure, T>> Function(Error error, StackTrace stacktrace)?
+        onError,
   }) async {
     final extendedExceptionConverters =
         _extendedExceptionConverters<T>(exceptionConverters);
 
-    return extendedExceptionConverters
-        .fold<FutureOr<Either<Failure, T>> Function(MessageLog? messageLog)>(
-      (messageLog) async {
-        final result = await task(messageLog);
+    try {
+      return await extendedExceptionConverters
+          .fold<FutureOr<Either<Failure, T>> Function(MessageLog? messageLog)>(
+        (messageLog) async {
+          final result = await task(messageLog);
 
-        if (messageLog != null) {
-          if (printOutput) messageLog.data['__output__'] = result;
+          if (messageLog != null) {
+            if (printOutput) messageLog.data['__output__'] = result;
 
-          result.fold(
-            (l) => logger.warn(messageLog),
-            (r) => logger.info(messageLog),
-          );
-        }
+            result.fold(
+              (l) => logger.warn(messageLog),
+              (r) => logger.info(messageLog),
+            );
+          }
 
-        return result;
-      },
-      (previousValue, element) => (messageLog) => element.observe(
-            task: previousValue,
-            logger: logger,
-            messageLog: messageLog,
-            printOutput: printOutput,
-          ),
-    )(messageLog);
+          return result;
+        },
+        (task, element) => (messageLog) => element.observe(
+              task: task,
+              logger: logger,
+              messageLog: messageLog,
+              printOutput: printOutput,
+            ),
+      )(messageLog);
+    }
+    // ignore: avoid_catching_errors
+    on Error catch (error, stacktrace) {
+      if (onError != null) return onError(error, stacktrace);
+      rethrow;
+    }
   }
 
-  /// Converts the given [error] into a [Failure].
+  /// Converts the given [exception] into a [Failure].
   ///
-  /// The [error] is expected to be an [Exception] object.
-  ///
-  /// If the given [error] is an [Error], then it is re-thrown. For more info,
-  /// see https://stackoverflow.com/a/57004304.
-  ///
-  /// If the given [error] is not an [Exception], then an [ArgumentError] is
-  /// thrown.
+  /// The [exception] is expected to be an [Exception] object.
   Failure convert({
-    required Object error,
+    required Exception exception,
     List<ExceptionConverter<Exception, void>>? exceptionConverters,
   }) {
-    if (error is! Exception) {
-      throw ArgumentError(
-        'The given error is not an Exception: ${error.runtimeType}',
-      );
-    }
-
     final extendedExceptionConverters =
         _extendedExceptionConverters(exceptionConverters);
 
     for (final exceptionConverter in extendedExceptionConverters) {
-      if (exceptionConverter.exceptionEquals(error)) {
-        return exceptionConverter.convert(error);
+      if (exceptionConverter.exceptionEquals(exception)) {
+        return exceptionConverter.convert(exception);
       }
     }
 
     // coverage:ignore-start
     throw StateError(
-      'No matching exception conversion found for ${error.runtimeType}',
+      'No matching exception conversion found for ${exception.runtimeType}',
     );
     // coverage:ignore-end
   }